Mailer test implemented
When a new user registers, successfully send a test email to my personal Gmail account. Important to note that emails are coming from an alternate personal Gmail account and this is using the email address and password stored in Mail.php. Need to look at a more secure method for storing sender details. Sender email and password are redacted from the commit.