Modified User Creation
When a BRAND registers (no longer referred to as 'advertiser' - this change in progress) their brand/company name is now stored in the 'username' column in the users database. Retiring the use of the 'company' column in the users database as it is pointless.

Other changes related to 'advertisers' changing to 'brands' included.